requirements.txt
fastapi==0.110.0
httpx==0.27.2
python-dotenv==1.0.1
redis==5.2.1
uvicorn==0.27.1
